feat: pss-ctr-ses-root-contract-008 - Close nested implementation re-exports from Sessions root
