+++ pkg/services/factory_sessions/contracts.go
@@ -2,10 +2,11 @@ package factorysessions
 
 import (
 	"context"
+	"io"
+	"io/fs"
 
 	interfaces "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
 	factoryruntime "github.com/portpowered/infinite-you/pkg/services/factory_runtime"
-	internalcontracts "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/contracts"
 	"github.com/portpowered/infinite-you/pkg/services/models"
 	operatorsettings "github.com/portpowered/infinite-you/pkg/services/operator_settings"
 	providersessions "github.com/portpowered/infinite-you/pkg/services/provider_sessions"
@@ -16,17 +17,49 @@ import (
 
 // InvocationMetric records one emitted runtime counter together with its
 // low-cardinality dimensions.
-type InvocationMetric = internalcontracts.InvocationMetric
+type InvocationMetric struct {
+	Name   string
+	Labels map[string]string
+}
 
-type (
-	ExecutionOpeningFileSystem           = internalcontracts.ExecutionOpeningFileSystem
-	DirectoryInspection                  = internalcontracts.DirectoryInspection
-	CursorPersistenceFileSystem          = internalcontracts.CursorPersistenceFileSystem
-	CursorPersistenceTemporaryFile       = internalcontracts.CursorPersistenceTemporaryFile
-	CursorPersistenceCreateTemporaryFile = internalcontracts.CursorPersistenceCreateTemporaryFile
-	RuntimePersistenceFileSystem         = internalcontracts.RuntimePersistenceFileSystem
-	InvocationMetricsRecorder            = internalcontracts.InvocationMetricsRecorder
-)
+// Effect-port contracts owned by the Sessions root. Nested internal/contracts
+// aliases these symbols for private implementation packages.
+type ExecutionOpeningFileSystem interface {
+	Getwd() (string, error)
+	Stat(string) (fs.FileInfo, error)
+}
+
+type DirectoryInspection interface {
+	Stat(string) (fs.FileInfo, error)
+	ReadDir(string) ([]fs.DirEntry, error)
+}
+
+type CursorPersistenceFileSystem interface {
+	MkdirAll(string, fs.FileMode) error
+	ReadFile(string) ([]byte, error)
+	Remove(string) error
+	Rename(string, string) error
+}
+
+type CursorPersistenceTemporaryFile interface {
+	io.Writer
+	Name() string
+	Chmod(fs.FileMode) error
+	Sync() error
+	Close() error
+}
+
+type CursorPersistenceCreateTemporaryFile func(string, string) (CursorPersistenceTemporaryFile, error)
+
+type RuntimePersistenceFileSystem interface {
+	MkdirAll(string, fs.FileMode) error
+	ReadFile(string) ([]byte, error)
+	WriteFile(string, []byte, fs.FileMode) error
+}
+
+type InvocationMetricsRecorder interface {
+	RecordInvocationMetric(InvocationMetric)
+}
 
 // InvocationTarget contains the detached configuration selected for one
 // invocation. Operations remain consumer-owned interfaces.
