+++ pkg/services/factory_sessions/internal/contracts/effects.go
@@ -1,43 +1,11 @@
 package contracts
 
-import (
-	"io"
-	"io/fs"
-)
+import factorysessions "github.com/portpowered/infinite-you/pkg/services/factory_sessions"
 
-type ExecutionOpeningFileSystem interface {
-	Getwd() (string, error)
-	Stat(string) (fs.FileInfo, error)
-}
-
-type DirectoryInspection interface {
-	Stat(string) (fs.FileInfo, error)
-	ReadDir(string) ([]fs.DirEntry, error)
-}
-
-type CursorPersistenceFileSystem interface {
-	MkdirAll(string, fs.FileMode) error
-	ReadFile(string) ([]byte, error)
-	Remove(string) error
-	Rename(string, string) error
-}
-
-type CursorPersistenceTemporaryFile interface {
-	io.Writer
-	Name() string
-	Chmod(fs.FileMode) error
-	Sync() error
-	Close() error
-}
-
-type CursorPersistenceCreateTemporaryFile func(string, string) (CursorPersistenceTemporaryFile, error)
-
-type RuntimePersistenceFileSystem interface {
-	MkdirAll(string, fs.FileMode) error
-	ReadFile(string) ([]byte, error)
-	WriteFile(string, []byte, fs.FileMode) error
-}
-
-type InvocationMetricsRecorder interface {
-	RecordInvocationMetric(InvocationMetric)
-}
+type ExecutionOpeningFileSystem = factorysessions.ExecutionOpeningFileSystem
+type DirectoryInspection = factorysessions.DirectoryInspection
+type CursorPersistenceFileSystem = factorysessions.CursorPersistenceFileSystem
+type CursorPersistenceTemporaryFile = factorysessions.CursorPersistenceTemporaryFile
+type CursorPersistenceCreateTemporaryFile = factorysessions.CursorPersistenceCreateTemporaryFile
+type RuntimePersistenceFileSystem = factorysessions.RuntimePersistenceFileSystem
+type InvocationMetricsRecorder = factorysessions.InvocationMetricsRecorder
