+++ pkg/services/factory_sessions/internal/contracts/identity.go
@@ -1,13 +1,13 @@
 // Package contracts owns implementation-facing Factory Sessions capability
-// types that must be shared across private implementation packages.
+// aliases shared across private implementation packages. Type ownership lives
+// on the Factory Sessions root.
 package contracts
 
+import factorysessions "github.com/portpowered/infinite-you/pkg/services/factory_sessions"
+
 // SessionIDGenerator supplies one opaque Factory Session identity.
-type SessionIDGenerator func() string
+type SessionIDGenerator = factorysessions.SessionIDGenerator
 
 // InvocationMetric records one emitted runtime counter together with its
 // low-cardinality dimensions.
-type InvocationMetric struct {
-	Name   string
-	Labels map[string]string
-}
+type InvocationMetric = factorysessions.InvocationMetric
