+++ pkg/services/factory_sessions/internal/execution/projection_consistency.go
@@ -6,6 +6,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	factorysessions "github.com/portpowered/infinite-you/pkg/services/factory_sessions"
 	"strings"
 	"time"
 
@@ -227,56 +228,6 @@ func ValidateResultMatchesEventProjection(result ResultReadResult, events []json
 	return nil
 }
 
-// ResultStatus is the customer-visible durable session result availability.
-type ResultStatus string
-
-const (
-	ResultStatusNotReady          ResultStatus = "NOT_READY"
-	ResultStatusPartial           ResultStatus = "PARTIAL"
-	ResultStatusFinal             ResultStatus = "FINAL"
-	ResultStatusFailedWithPartial ResultStatus = "FAILED_WITH_PARTIAL"
-	ResultStatusUnavailable       ResultStatus = "UNAVAILABLE"
-)
-
-// ResultMode selects final or partial durable result retrieval.
-type ResultMode string
-
-const (
-	ResultModeFinal   ResultMode = "final"
-	ResultModePartial ResultMode = "partial"
-)
-
-// ResultRequest normalizes durable session result read parameters.
-type ResultRequest struct {
-	Mode             ResultMode
-	IncludeArtifacts bool
-}
-
-// ResultAvailabilityDetail explains why a durable result is not ready or unavailable.
-type ResultAvailabilityDetail struct {
-	Reason    string
-	Message   string
-	Retryable bool
-}
-
-// ResultReadResult is the shared durable session result projection consumed by API,
-// CLI, MCP, and UI transports.
-type ResultReadResult struct {
-	SessionID        string
-	ResultStatus     ResultStatus
-	SessionStatus    LifecycleStatus
-	Mode             ResultMode
-	IncludeArtifacts bool
-	PrimaryResult    json.RawMessage
-	ArtifactIDs      []string
-	ArtifactRefs     []ArtifactRefSummary
-	Failure          *FailureSummary
-	Availability     *ResultAvailabilityDetail
-}
-
-// DispatchStatus is the canonical dispatch lifecycle status shared across orchestrators.
-type DispatchStatus string
-
 const (
 	DispatchStatusQueued      DispatchStatus = "QUEUED"
 	DispatchStatusRunning     DispatchStatus = "RUNNING"
@@ -287,103 +238,6 @@ const (
 	DispatchStatusSkipped     DispatchStatus = "SKIPPED"
 	DispatchStatusInterrupted DispatchStatus = "INTERRUPTED"
 )
-
-// DispatchUsage summarizes one dispatch execution.
-type DispatchUsage struct {
-	InputTokens    int64
-	OutputTokens   int64
-	TotalTokens    int64
-	DurationMillis int64
-	CostUSD        float64
-	RetryCount     int32
-}
-
-// DispatchWarning exposes one customer-visible dispatch warning.
-type DispatchWarning struct {
-	Code    string
-	Message string
-}
-
-// DispatchFailureDetail exposes one customer-visible dispatch failure.
-type DispatchFailureDetail struct {
-	Reason  string
-	Message string
-}
-
-// DispatchPetriProjection carries Petri-specific dispatch metadata.
-type DispatchPetriProjection struct {
-	TransitionID    string
-	WorkstationName string
-	WorkerType      string
-}
-
-// DispatchJavaScriptProjection carries JavaScript-specific dispatch metadata.
-type DispatchJavaScriptProjection struct {
-	TaskKind      string
-	TaskLabel     string
-	ExecutionMode string
-}
-
-// ProviderSessionRef correlates one dispatch with a provider session identity.
-type ProviderSessionRef struct {
-	Provider string
-	Kind     string
-	ID       string
-}
-
-// DispatchSummary is the shared durable dispatch list projection.
-type DispatchSummary struct {
-	ID                    string
-	Status                DispatchStatus
-	DispatchKind          string
-	Phase                 string
-	Label                 string
-	Attempt               int
-	Retryable             *bool
-	FailureClassification string
-	RunnerID              string
-	PresetID              string
-	ModelProvider         string
-	Model                 string
-	ReasoningEffort       string
-	Provider              string
-	ProviderSessionRefs   []ProviderSessionRef
-	OutputArtifactIDs     []string
-	Usage                 *DispatchUsage
-	Warnings              []DispatchWarning
-	FailureDetail         *DispatchFailureDetail
-	JavaScript            *DispatchJavaScriptProjection
-}
-
-// DispatchDetail is the shared durable dispatch read projection.
-type DispatchDetail struct {
-	DispatchSummary
-	SessionID         string
-	OrchestratorKind  string
-	ArtifactIDs       []string
-	StatusTransitions []DispatchStatus
-	Petri             *DispatchPetriProjection
-	JavaScript        *DispatchJavaScriptProjection
-}
-
-// ListDispatchesResult is the shared durable dispatch list outcome.
-type ListDispatchesResult struct {
-	SessionID  string
-	Dispatches []DispatchSummary
-}
-
-// DispatchFilters narrows a canonical Dispatch read by phase and status.
-type DispatchFilters struct {
-	Phase  string
-	Status DispatchStatus
-}
-
-// DispatchQueryRequest is the complete owner request for one filtered
-// Factory Session Dispatch inventory read.
-type DispatchQueryRequest struct {
-	SessionID string
-	Filters   DispatchFilters
-}
 
 // NormalizeDispatchFilters validates and canonicalizes transport-provided filters.
 func NormalizeDispatchFilters(filters DispatchFilters) (DispatchFilters, error) {
@@ -441,80 +295,6 @@ func isCanonicalDispatchStatus(status DispatchStatus) bool {
 	default:
 		return false
 	}
-}
-
-// ArtifactRetrievalRef is a safe API-relative artifact retrieval reference.
-type ArtifactRetrievalRef struct {
-	Href   string
-	Method string
-}
-
-// ArtifactRedactionCounts summarizes secret suppression for one artifact.
-type ArtifactRedactionCounts struct {
-	Paths   int32
-	Secrets int32
-	Tokens  int32
-}
-
-// ArtifactSummary is the shared durable artifact list projection.
-type ArtifactSummary struct {
-	ID              string
-	Kind            string
-	Visibility      string
-	Label           string
-	ContentHash     string
-	SizeBytes       int64
-	CreatedAt       *time.Time
-	DispatchID      string
-	AuditMode       string
-	RedactionCounts *ArtifactRedactionCounts
-	RetrievalRef    *ArtifactRetrievalRef
-}
-
-// ArtifactDetail is the shared durable artifact read projection.
-type ArtifactDetail struct {
-	ArtifactSummary
-	SessionID       string
-	Summary         string
-	CaptureMetadata map[string]any
-	Content         json.RawMessage
-	ContentRef      *ArtifactRetrievalRef
-}
-
-// ListArtifactsResult is the shared durable artifact list outcome.
-type ListArtifactsResult struct {
-	SessionID string
-	Artifacts []ArtifactSummary
-}
-
-// EventReconnectRequest identifies the last acknowledged durable session event.
-type EventReconnectRequest struct {
-	AfterEventID  string
-	AfterSequence *int
-}
-
-// EventReadResult carries replayed canonical session events for one durable session.
-type EventReadResult struct {
-	SessionID string
-	Events    []json.RawMessage
-}
-
-// MaterializeEventReadStream owns the finite stream lifecycle for one durable
-// event read. Transports receive an already-closed live channel plus detached
-// canonical history and do not manufacture channel-backed streams.
-func MaterializeEventReadStream(result EventReadResult) *interfaces.FactoryEventStream {
-	closed := make(chan interfaces.FactoryEvent)
-	close(closed)
-	stream := &interfaces.FactoryEventStream{Events: closed}
-	for _, raw := range result.Events {
-		var event interfaces.FactoryEvent
-		if err := json.Unmarshal(raw, &event); err != nil {
-			continue
-		}
-		event.Payload = append(json.RawMessage(nil), event.Payload...)
-		stream.History = append(stream.History, event)
-	}
-	return stream
 }
 
 // NormalizeResultRequest validates and normalizes one durable result read request.
@@ -1019,4 +799,8 @@ func stringValuePtr(value *string) string {
 func timePtr(value time.Time) *time.Time {
 	cloned := value
 	return &cloned
+}
+
+func MaterializeEventReadStream(result EventReadResult) *interfaces.FactoryEventStream {
+	return factorysessions.MaterializeEventReadStream(result)
 }
